requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy[asyncio]
alembic
asyncpg
pydantic-settings
apscheduler
python-dateutil